Identify the edges of the paper of each page

Identify the edges of the paper of each page

paper_shape(data)

Arguments

data
data.frame created by ggpage_build.

Value

data.frame,

Examples

paper_shape(ggpage_build(tinderbox))

#> # A tibble: 9 x 5
#>    page  xmin  xmax  ymin  ymax
#>   <int> <dbl> <dbl> <dbl> <dbl>
#> 1     1   168    90  -114  -213
#> 2     2   167    90  -224  -323
#> 3     3   168    90  -334  -433
#> 4     4   259   180  -114  -213
#> 5     5   258   180  -224  -323
#> 6     6   259   180  -334  -433
#> 7     7   349   270  -114  -213
#> 8     8   349   270  -224  -323
#> 9     9   348   270  -334  -377